Packaging project as multiple files, note: performance issues to be resolved
